Converted entire project to static library; Fixed warnings

+++ LSAP/src/LSAP/ImGui/LSGui.cpp
@@ -74,7 +74,7 @@ namespace LSAP {
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::getApplication();
-		io.DisplaySize = ImVec2(app.getWindow().getWidth(), app.getWindow().getHeight());
+		io.DisplaySize = ImVec2((float)app.getWindow().getWidth(), (float)app.getWindow().getHeight());
 
 		// Rendering
 		ImGui::Render();
